+++ core/vm/interface.go
@@ -39,7 +39,7 @@ type StateDB interface {
 	GetPoseidonCodeHash(common.Address) common.Hash
 	GetCode(common.Address) []byte
 	SetCode(common.Address, []byte)
-	GetCodeSize(common.Address) int
+	GetCodeSize(common.Address) uint64
 
 	AddRefund(uint64)
 	SubRefund(uint64)
